+++ src/Oro/Bundle/WorkflowBundle/CHANGELOG.md
@@ -40,3 +40,10 @@ CHANGELOG for BAP-10514
 * Added new method `hasActiveWorkflowsByEntityClass`, that indicates if an entity class has one or more linked workflows
 * Removed method `getActiveWorkflowByEntityClass` from `Oro\Bundle\WorkflowBundle\Model\WorkflowRegistry`, use `getActiveWorkflowsByEntityClass`
 * Removed method `hasActiveWorkflowByEntityClass` from `Oro\Bundle\WorkflowBundle\Model\WorkflowRegistry`, use `hasActiveWorkflowsByEntityClass`
+
+CHANGELOG for BAP-11201
+=======================
+* Removed service `@oro_workflow.manager.system_config` and its class `Oro\Bundle\WorkflowBundle\Model\WorkflowSystemConfigManager` as now there no entity configuration for active workflows. Activation and deactivation of a workflow now should be managed through WorkflowManager (`Oro\Bundle\WorkflowBundle\Model\WorkflowManager`- `@@oro_workflow.manager`)
+ 
+
+